pyproject.toml
[project]
name = "recipe-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.0.0",
    "fastapi>=0.116.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
